requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
opencv-python>=4.9.0.80
numpy>=1.26.0
pydantic>=2.6.0
httpx>=0.27.0
langchain>=0.3.0
langchain-mcp-adapters>=0.1.0
fastmcp>=2.0.0
chromadb>=1.0.0
google-genai>=1.0.0
